Optimize jwt guard and extract wsjwt guard
